Fix review issues
