+++ pallets/multicollateral-bonding-curve-pool/src/lib.rs
@@ -224,7 +224,7 @@ pub mod pallet {
     }
 
     /// The current storage version.
-    const STORAGE_VERSION: StorageVersion = StorageVersion::new(3);
+    const STORAGE_VERSION: StorageVersion = StorageVersion::new(4);
 
     #[pallet::pallet]
     #[pallet::generate_store(pub(super) trait Store)]
